switch back to resampling
